+++ src/main/java/com/inditex/hiring/infrastructure/persistence/jpa/write/OfferEntity.java
@@ -16,7 +16,7 @@ public class OfferEntity {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
     @Column(name = "BRAND_ID")
-    private Long brandId;
+    private Integer brandId;
     @Column(name = "PRICE_LIST")
     private Long priceListId;
     @Column(name = "START_DATE")
@@ -24,7 +24,7 @@ public class OfferEntity {
     @Column(name = "END_DATE")
     private Instant endDate;
     @Column(name = "PRIORITY")
-    private Long priority;
+    private Integer priority;
     @Column(name = "PARTNUMBER")
     private String partNumber;
     @Column(name = "PRICE")
