+++ src/test/java/com/inditex/hiring/e2e/offer/models/request/AddOfferRequest.java
@@ -1,5 +1,5 @@
 package com.inditex.hiring.e2e.offer.models.request;
 
 public record AddOfferRequest(int brandId, long priceListId, String startDate,
-                              String endDate, long priority, String partNumber, double price, String currencyISO) {
+                              String endDate, int priority, String partNumber, double price, String currencyISO) {
 }
